requirements.txt
robotframework>=3.2.2
robotframework-pythonlibcore>=3.0.0
